requirements.txt
python-docx==1.2.0
lxml==4.9.3
Flask==2.3.3
Werkzeug==2.3.7